+++ src/document/crdt/root.ts
@@ -20,32 +20,32 @@ import {
   TimeTicket,
 } from '@yorkie-js-sdk/src/document/time/ticket';
 import {
-  JSONContainer,
-  JSONElement,
-  TextElement,
-} from '@yorkie-js-sdk/src/document/json/element';
-import { ObjectInternal } from '@yorkie-js-sdk/src/document/json/object';
+  CRDTContainer,
+  CRDTElement,
+  CRDTTextElement,
+} from '@yorkie-js-sdk/src/document/crdt/element';
+import { CRDTObject } from '@yorkie-js-sdk/src/document/crdt/object';
 
-interface JSONElementPair {
-  element: JSONElement;
-  parent?: JSONContainer;
+interface CRDTElementPair {
+  element: CRDTElement;
+  parent?: CRDTContainer;
 }
 
 /**
- * `JSONRoot` is a structure represents the root of JSON. It has a hash table of
- * all JSON elements to find a specific element when applying remote changes
+ * `CRDTRoot` is a structure represents the root. It has a hash table of
+ * all elements to find a specific element when applying remote changes
  * received from server.
  *
  * Every element has a unique time ticket at creation, which allows us to find
  * a particular element.
  */
-export class JSONRoot {
-  private rootObject: ObjectInternal;
-  private elementPairMapByCreatedAt: Map<string, JSONElementPair>;
+export class CRDTRoot {
+  private rootObject: CRDTObject;
+  private elementPairMapByCreatedAt: Map<string, CRDTElementPair>;
   private removedElementSetByCreatedAt: Set<string>;
   private textWithGarbageSetByCreatedAt: Set<string>;
 
-  constructor(rootObject: ObjectInternal) {
+  constructor(rootObject: CRDTObject) {
     this.rootObject = rootObject;
     this.elementPairMapByCreatedAt = new Map();
     this.removedElementSetByCreatedAt = new Set();
@@ -57,7 +57,7 @@ export class JSONRoot {
     );
 
     rootObject.getDescendants(
-      (elem: JSONElement, parent: JSONContainer): boolean => {
+      (elem: CRDTElement, parent: CRDTContainer): boolean => {
         this.registerElement(elem, parent);
         return false;
       },
@@ -67,14 +67,14 @@ export class JSONRoot {
   /**
    * `create` creates a new instance of Root.
    */
-  public static create(): JSONRoot {
-    return new JSONRoot(ObjectInternal.create(InitialTimeTicket));
+  public static create(): CRDTRoot {
+    return new CRDTRoot(CRDTObject.create(InitialTimeTicket));
   }
 
   /**
    * `findByCreatedAt` returns the element of given creation time.
    */
-  public findByCreatedAt(createdAt: TimeTicket): JSONElement | undefined {
+  public findByCreatedAt(createdAt: TimeTicket): CRDTElement | undefined {
     const pair = this.elementPairMapByCreatedAt.get(createdAt.toIDString());
     if (!pair) {
       return;
@@ -115,7 +115,7 @@ export class JSONRoot {
   /**
    * `registerElement` registers the given element to hash table.
    */
-  public registerElement(element: JSONElement, parent: JSONContainer): void {
+  public registerElement(element: CRDTElement, parent: CRDTContainer): void {
     this.elementPairMapByCreatedAt.set(element.getCreatedAt().toIDString(), {
       parent,
       element,
@@ -125,7 +125,7 @@ export class JSONRoot {
   /**
    * `deregisterElement` deregister the given element from hash table.
    */
-  public deregisterElement(element: JSONElement): void {
+  public deregisterElement(element: CRDTElement): void {
     this.elementPairMapByCreatedAt.delete(element.getCreatedAt().toIDString());
     this.removedElementSetByCreatedAt.delete(
       element.getCreatedAt().toIDString(),
@@ -135,14 +135,14 @@ export class JSONRoot {
   /**
    * `registerRemovedElement` registers the given element to hash table.
    */
-  public registerRemovedElement(element: JSONElement): void {
+  public registerRemovedElement(element: CRDTElement): void {
     this.removedElementSetByCreatedAt.add(element.getCreatedAt().toIDString());
   }
 
   /**
    * `registerTextWithGarbage` registers the given text to hash set.
    */
-  public registerTextWithGarbage(text: TextElement): void {
+  public registerTextWithGarbage(text: CRDTTextElement): void {
     this.textWithGarbageSetByCreatedAt.add(text.getCreatedAt().toIDString());
   }
 
@@ -163,7 +163,7 @@ export class JSONRoot {
   /**
    * `getObject` returns root object.
    */
-  public getObject(): ObjectInternal {
+  public getObject(): CRDTObject {
     return this.rootObject;
   }
 
@@ -176,7 +176,7 @@ export class JSONRoot {
     for (const createdAt of this.removedElementSetByCreatedAt) {
       count++;
       const pair = this.elementPairMapByCreatedAt.get(createdAt)!;
-      if (pair.element instanceof JSONContainer) {
+      if (pair.element instanceof CRDTContainer) {
         pair.element.getDescendants(() => {
           count++;
           return false;
@@ -186,7 +186,7 @@ export class JSONRoot {
 
     for (const createdAt of this.textWithGarbageSetByCreatedAt) {
       const pair = this.elementPairMapByCreatedAt.get(createdAt)!;
-      const text = pair.element as TextElement;
+      const text = pair.element as CRDTTextElement;
       count += text.getRemovedNodesLen();
     }
 
@@ -196,8 +196,8 @@ export class JSONRoot {
   /**
    * `deepcopy` copies itself deeply.
    */
-  public deepcopy(): JSONRoot {
-    return new JSONRoot(this.rootObject.deepcopy());
+  public deepcopy(): CRDTRoot {
+    return new CRDTRoot(this.rootObject.deepcopy());
   }
 
   /**
@@ -219,7 +219,7 @@ export class JSONRoot {
 
     for (const createdAt of this.textWithGarbageSetByCreatedAt) {
       const pair = this.elementPairMapByCreatedAt.get(createdAt)!;
-      const text = pair.element as TextElement;
+      const text = pair.element as CRDTTextElement;
 
       const removedNodeCnt = text.purgeTextNodesWithGarbage(ticket);
       if (removedNodeCnt > 0) {
@@ -233,11 +233,11 @@ export class JSONRoot {
     return count;
   }
 
-  private garbageCollectInternal(element: JSONElement): number {
+  private garbageCollectInternal(element: CRDTElement): number {
     let count = 0;
 
     // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    const callback = (elem: JSONElement, parent?: JSONContainer): boolean => {
+    const callback = (elem: CRDTElement, parent?: CRDTContainer): boolean => {
       this.deregisterElement(elem);
       count++;
       return false;
@@ -245,7 +245,7 @@ export class JSONRoot {
 
     callback(element);
 
-    if (element instanceof JSONContainer) {
+    if (element instanceof CRDTContainer) {
       element.getDescendants(callback);
     }
