+++ src/document/operation/edit_operation.ts
@@ -16,9 +16,9 @@
 
 import { logger } from '@yorkie-js-sdk/src/util/logger';
 import { TimeTicket } from '@yorkie-js-sdk/src/document/time/ticket';
-import { JSONRoot } from '@yorkie-js-sdk/src/document/json/root';
-import { RGATreeSplitNodePos } from '@yorkie-js-sdk/src/document/json/rga_tree_split';
-import { PlainTextInternal } from '@yorkie-js-sdk/src/document/json/plain_text';
+import { CRDTRoot } from '@yorkie-js-sdk/src/document/crdt/root';
+import { RGATreeSplitNodePos } from '@yorkie-js-sdk/src/document/crdt/rga_tree_split';
+import { CRDTPlainText } from '@yorkie-js-sdk/src/document/crdt/plain_text';
 import { Operation } from '@yorkie-js-sdk/src/document/operation/operation';
 
 /**
@@ -69,11 +69,11 @@ export class EditOperation extends Operation {
   /**
    * `execute` executes this operation on the given document(`root`).
    */
-  public execute(root: JSONRoot): void {
+  public execute(root: CRDTRoot): void {
     const parentObject = root.findByCreatedAt(this.getParentCreatedAt());
-    if (parentObject instanceof PlainTextInternal) {
-      const text = parentObject as PlainTextInternal;
-      text.editInternal(
+    if (parentObject instanceof CRDTPlainText) {
+      const text = parentObject as CRDTPlainText;
+      text.edit(
         [this.fromPos, this.toPos],
         this.content,
         this.getExecutedAt(),
