+++ src/document/operation/select_operation.ts
@@ -16,10 +16,10 @@
 
 import { logger } from '@yorkie-js-sdk/src/util/logger';
 import { TimeTicket } from '@yorkie-js-sdk/src/document/time/ticket';
-import { JSONRoot } from '@yorkie-js-sdk/src/document/json/root';
-import { RGATreeSplitNodePos } from '@yorkie-js-sdk/src/document/json/rga_tree_split';
-import { PlainTextInternal } from '@yorkie-js-sdk/src/document/json/plain_text';
-import { RichTextInternal } from '@yorkie-js-sdk/src/document/json/rich_text';
+import { CRDTRoot } from '@yorkie-js-sdk/src/document/crdt/root';
+import { RGATreeSplitNodePos } from '@yorkie-js-sdk/src/document/crdt/rga_tree_split';
+import { CRDTPlainText } from '@yorkie-js-sdk/src/document/crdt/plain_text';
+import { CRDTRichText } from '@yorkie-js-sdk/src/document/crdt/rich_text';
 import { Operation } from '@yorkie-js-sdk/src/document/operation/operation';
 
 /**
@@ -55,14 +55,14 @@ export class SelectOperation extends Operation {
   /**
    * `execute` executes this operation on the given document(`root`).
    */
-  public execute(root: JSONRoot): void {
+  public execute(root: CRDTRoot): void {
     const parentObject = root.findByCreatedAt(this.getParentCreatedAt());
-    if (parentObject instanceof PlainTextInternal) {
-      const text = parentObject as PlainTextInternal;
-      text.selectInternal([this.fromPos, this.toPos], this.getExecutedAt());
-    } else if (parentObject instanceof RichTextInternal) {
-      const text = parentObject as RichTextInternal;
-      text.selectInternal([this.fromPos, this.toPos], this.getExecutedAt());
+    if (parentObject instanceof CRDTPlainText) {
+      const text = parentObject as CRDTPlainText;
+      text.select([this.fromPos, this.toPos], this.getExecutedAt());
+    } else if (parentObject instanceof CRDTRichText) {
+      const text = parentObject as CRDTRichText;
+      text.select([this.fromPos, this.toPos], this.getExecutedAt());
     } else {
       if (!parentObject) {
         logger.fatal(`fail to find ${this.getParentCreatedAt()}`);
