+++ test/integration/array_test.ts
@@ -3,7 +3,12 @@ import { DocumentReplica } from '@yorkie-js-sdk/src/document/document';
 import { MaxTimeTicket } from '@yorkie-js-sdk/src/document/time/ticket';
 
 import { withTwoClientsAndDocuments } from '@yorkie-js-sdk/test/integration/integration_helper';
-import { JSONArray, JSONElement, TimeTicket } from '@yorkie-js-sdk/src/yorkie';
+import {
+  JSONArray,
+  WrappedElement,
+  Primitive,
+  TimeTicket,
+} from '@yorkie-js-sdk/src/yorkie';
 
 describe('Array', function () {
   it('should handle delete operations', function () {
@@ -11,13 +16,13 @@ describe('Array', function () {
     assert.equal('{}', doc.toSortedJSON());
 
     doc.update((root) => {
-      root['k1'] = ['1', '2', '3'];
+      root.k1 = ['1', '2', '3'];
     }, 'set {"k1":["1","2","3"]}');
     assert.equal('{"k1":["1","2","3"]}', doc.toSortedJSON());
 
     doc.update((root) => {
-      delete root['k1'][1];
-      root['k1'].push('4');
+      delete root.k1[1];
+      root.k1.push('4');
     }, 'set {"k1":["1","3","4"]}');
     assert.equal('{"k1":["1","3","4"]}', doc.toSortedJSON());
   });
@@ -30,17 +35,17 @@ describe('Array', function () {
     assert.equal('{}', doc.toSortedJSON());
 
     doc.update((root) => {
-      root['k1'] = ['1', '2', '3'];
+      root.k1 = ['1', '2', '3'];
     }, 'set {"k1":["1","2","3"]}');
     assert.equal('{"k1":["1","2","3"]}', doc.toSortedJSON());
 
     doc.update((root) => {
-      delete root['k1'][1];
-      root['k1'].push('4');
+      delete root.k1[1];
+      root.k1.push('4');
     }, 'set {"k1":["1","3","4"]}');
 
     doc.update((root) => {
-      root['k1'].push([4, 5, 6]);
+      root.k1.push([4, 5, 6]);
       assert.equal('{"k1":["1","3","4",[4,5,6]]}', root.toJSON!());
     });
 
@@ -54,17 +59,17 @@ describe('Array', function () {
     assert.equal('{}', doc.toSortedJSON());
 
     doc.update((root) => {
-      root['k1'] = ['1', '2', '3'];
+      root.k1 = ['1', '2', '3'];
     }, 'set {"k1":["1","2","3"]}');
     assert.equal('{"k1":["1","2","3"]}', doc.toSortedJSON());
 
     doc.update((root) => {
-      delete root['k1'][1];
-      root['k1'].push('4');
+      delete root.k1[1];
+      root.k1.push('4');
     }, 'set {"k1":["1","3","4"]}');
 
     doc.update((root) => {
-      root['k1'].push({ a: '1', b: '2' });
+      root.k1.push({ a: '1', b: '2' });
       assert.equal('{"k1":["1","3","4",{"a":"1","b":"2"}]}', root.toJSON!());
     });
 
@@ -89,25 +94,25 @@ describe('Array', function () {
     const doc = DocumentReplica.create<{ list: JSONArray<number> }>('test-doc');
     assert.equal('{}', doc.toSortedJSON());
 
-    let toDelete: JSONElement;
+    let target: WrappedElement;
     doc.update((root) => {
-      root['list'] = [];
-      assert.equal(1, root['list'].push(4));
-      assert.equal(2, root['list'].push(3));
-      assert.equal(3, root['list'].push(2));
-      assert.equal(4, root['list'].push(1));
-      toDelete = root['list'].getElementByIndex!(2);
+      root.list = [];
+      assert.equal(1, root.list.push(4));
+      assert.equal(2, root.list.push(3));
+      assert.equal(3, root.list.push(2));
+      assert.equal(4, root.list.push(1));
+      target = root.list.getElementByIndex!(2);
     }, 'set {"list":[4,3,2,1]}');
 
     assert.equal('{"list":[4,3,2,1]}', doc.toSortedJSON());
 
     doc.update((root) => {
-      root['list'].deleteByID!(toDelete.getID());
+      root.list.deleteByID!(target.getID!());
     }, 'delete 2');
     assert.equal('{"list":[4,3,1]}', doc.toSortedJSON());
 
     doc.update((root) => {
-      assert.equal(4, root['list'].push(2));
+      assert.equal(4, root.list.push(2));
     }, 'push 2');
     assert.equal('{"list":[4,3,1,2]}', doc.toSortedJSON());
   });
@@ -116,36 +121,36 @@ describe('Array', function () {
     const doc = DocumentReplica.create<{ list: JSONArray<number> }>('test-doc');
     assert.equal('{}', doc.toSortedJSON());
 
-    let prev: JSONElement;
+    let prev: WrappedElement;
     doc.update((root) => {
-      root['list'] = [];
-      root['list'].push(1);
-      root['list'].push(2);
-      root['list'].push(4);
-      prev = root['list'].getElementByIndex!(1);
+      root.list = [];
+      root.list.push(1);
+      root.list.push(2);
+      root.list.push(4);
+      prev = root.list.getElementByIndex!(1);
     }, 'set {"list":[1,2,4]}');
 
     assert.equal('{"list":[1,2,4]}', doc.toSortedJSON());
 
     doc.update((root) => {
-      root['list'].insertAfter!(prev.getID(), 3);
+      root.list.insertAfter!(prev.getID!(), 3);
     }, 'insert 3');
     assert.equal('{"list":[1,2,3,4]}', doc.toSortedJSON());
 
     doc.update((root) => {
-      delete root['list'][1];
+      delete root.list[1];
     }, 'remove 2');
     assert.equal('{"list":[1,3,4]}', doc.toSortedJSON());
 
     doc.update((root) => {
-      prev = root['list'].getElementByIndex!(0);
-      root['list'].insertAfter!(prev.getID(), 2);
+      prev = root.list.getElementByIndex!(0);
+      root.list.insertAfter!(prev.getID!(), 2);
     }, 'insert 2');
     assert.equal('{"list":[1,2,3,4]}', doc.toSortedJSON());
 
     const root = doc.getRoot();
-    for (let idx = 0; idx < root['list'].length; idx++) {
-      assert.equal(idx + 1, root['list'][idx]);
+    for (let idx = 0; idx < root.list.length; idx++) {
+      assert.equal(idx + 1, root.list[idx]);
     }
   });
 
@@ -154,20 +159,20 @@ describe('Array', function () {
     assert.equal('{}', doc.toSortedJSON());
 
     doc.update((root) => {
-      root['list'] = [0, 1, 2];
+      root.list = [0, 1, 2];
     }, 'set {"list":[0,1,2]}');
 
     doc.update((root) => {
-      const next = root['list'].getElementByIndex!(0);
-      const item = root['list'].getElementByIndex!(2);
-      root['list'].moveBefore!(next.getID(), item.getID());
+      const next = root.list.getElementByIndex!(0);
+      const item = root.list.getElementByIndex!(2);
+      root.list.moveBefore!(next.getID!(), item.getID!());
       assert.equal('{"list":[2,0,1]}', root.toJSON!());
     });
 
     doc.update((root) => {
-      const next = root['list'].getElementByIndex!(0);
-      const item = root['list'].getElementByIndex!(2);
-      root['list'].moveBefore!(next.getID(), item.getID());
+      const next = root.list.getElementByIndex!(0);
+      const item = root.list.getElementByIndex!(2);
+      root.list.moveBefore!(next.getID!(), item.getID!());
       assert.equal('{"list":[1,2,0]}', root.toJSON!());
     });
   });
@@ -177,20 +182,20 @@ describe('Array', function () {
     assert.equal('{}', doc.toSortedJSON());
 
     doc.update((root) => {
-      root['list'] = [0, 1, 2];
+      root.list = [0, 1, 2];
     }, 'set {"list":[0,1,2]}');
 
     doc.update((root) => {
-      const prev = root['list'].getElementByIndex!(0);
-      const item = root['list'].getElementByIndex!(2);
-      root['list'].moveAfter!(prev.getID(), item.getID());
+      const prev = root.list.getElementByIndex!(0);
+      const item = root.list.getElementByIndex!(2);
+      root.list.moveAfter!(prev.getID!(), item.getID!());
       assert.equal('{"list":[0,2,1]}', root.toJSON!());
     });
 
     doc.update((root) => {
-      const prev = root['list'].getElementByIndex!(0);
-      const item = root['list'].getElementByIndex!(2);
-      root['list'].moveAfter!(prev.getID(), item.getID());
+      const prev = root.list.getElementByIndex!(0);
+      const item = root.list.getElementByIndex!(2);
+      root.list.moveAfter!(prev.getID!(), item.getID!());
       assert.equal('{"list":[0,1,2]}', root.toJSON!());
     });
   });
@@ -200,24 +205,24 @@ describe('Array', function () {
     assert.equal('{}', doc.toSortedJSON());
 
     doc.update((root) => {
-      root['list'] = [0, 1, 2];
+      root.list = [0, 1, 2];
     }, 'set {"list":[0,1,2]}');
 
     doc.update((root) => {
-      const item = root['list'].getElementByIndex!(2);
-      root['list'].moveFront!(item.getID());
+      const item = root.list.getElementByIndex!(2);
+      root.list.moveFront!(item.getID!());
       assert.equal('{"list":[2,0,1]}', root.toJSON!());
     });
 
     doc.update((root) => {
-      const item = root['list'].getElementByIndex!(1);
-      root['list'].moveFront!(item.getID());
+      const item = root.list.getElementByIndex!(1);
+      root.list.moveFront!(item.getID!());
       assert.equal('{"list":[0,2,1]}', root.toJSON!());
     });
 
     doc.update((root) => {
-      const item = root['list'].getElementByIndex!(0);
-      root['list'].moveFront!(item.getID());
+      const item = root.list.getElementByIndex!(0);
+      root.list.moveFront!(item.getID!());
       assert.equal('{"list":[0,2,1]}', root.toJSON!());
     });
   });
@@ -227,24 +232,24 @@ describe('Array', function () {
     assert.equal('{}', doc.toSortedJSON());
 
     doc.update((root) => {
-      root['list'] = [0, 1, 2];
+      root.list = [0, 1, 2];
     }, 'set {"list":[0,1,2]}');
 
     doc.update((root) => {
-      const item = root['list'].getElementByIndex!(2);
-      root['list'].moveLast!(item.getID());
+      const item = root.list.getElementByIndex!(2);
+      root.list.moveLast!(item.getID!());
       assert.equal('{"list":[0,1,2]}', root.toJSON!());
     });
 
     doc.update((root) => {
-      const item = root['list'].getElementByIndex!(1);
-      root['list'].moveLast!(item.getID());
+      const item = root.list.getElementByIndex!(1);
+      root.list.moveLast!(item.getID!());
       assert.equal('{"list":[0,2,1]}', root.toJSON!());
     });
 
     doc.update((root) => {
-      const item = root['list'].getElementByIndex!(0);
-      root['list'].moveLast!(item.getID());
+      const item = root.list.getElementByIndex!(0);
+      root.list.moveLast!(item.getID!());
       assert.equal('{"list":[2,1,0]}', root.toJSON!());
     });
   });
@@ -252,21 +257,21 @@ describe('Array', function () {
   it('Can handle concurrent insertAfter operations', async function () {
     type TestDoc = { k1: JSONArray<number> };
     await withTwoClientsAndDocuments<TestDoc>(async (c1, d1, c2, d2) => {
-      let prev: JSONElement;
+      let prev: WrappedElement;
       d1.update((root) => {
-        root['k1'] = [1, 2, 3, 4];
-        prev = root['k1'].getElementByIndex!(1);
+        root.k1 = [1, 2, 3, 4];
+        prev = root.k1.getElementByIndex!(1);
       });
       await c1.sync();
       await c2.sync();
       assert.equal(d1.toSortedJSON(), d2.toSortedJSON());
 
       d1.update((root) => {
-        root['k1'].deleteByID!(prev.getID());
+        root.k1.deleteByID!(prev.getID!());
         assert.equal('{"k1":[1,3,4]}', root.toJSON!());
       });
       d2.update((root) => {
-        root['k1'].insertAfter!(prev.getID(), 2);
+        root.k1.insertAfter!(prev.getID!(), 2);
         assert.equal('{"k1":[1,2,2,3,4]}', root.toJSON!());
       });
       await c1.sync();
@@ -276,13 +281,13 @@ describe('Array', function () {
       assert.equal('{"k1":[1,2,3,4]}', d1.toJSON!());
 
       d1.update((root) => {
-        const prev = root['k1'].getElementByIndex!(1);
-        root['k1'].insertAfter!(prev.getID(), '2.1');
+        const prev = root.k1.getElementByIndex!(1);
+        root.k1.insertAfter!(prev.getID!(), '2.1');
         assert.equal('{"k1":[1,2,"2.1",3,4]}', root.toJSON!());
       });
       d2.update((root) => {
-        const prev = root['k1'].getElementByIndex!(1);
-        root['k1'].insertAfter!(prev.getID(), '2.2');
+        const prev = root.k1.getElementByIndex!(1);
+        root.k1.insertAfter!(prev.getID!(), '2.2');
         assert.equal('{"k1":[1,2,"2.2",3,4]}', root.toJSON!());
       });
       await c1.sync();
@@ -296,7 +301,7 @@ describe('Array', function () {
     type TestDoc = { k1: JSONArray<number> };
     await withTwoClientsAndDocuments<TestDoc>(async (c1, d1, c2, d2) => {
       d1.update((root) => {
-        root['k1'] = [0, 1, 2];
+        root.k1 = [0, 1, 2];
         assert.equal('{"k1":[0,1,2]}', root.toJSON!());
       });
       await c1.sync();
@@ -304,30 +309,30 @@ describe('Array', function () {
       assert.equal(d1.toJSON!(), d2.toJSON!());
 
       d1.update((root) => {
-        const next = root['k1'].getElementByIndex!(0);
-        const item = root['k1'].getElementByIndex!(2);
-        root['k1'].moveBefore!(next.getID(), item.getID());
+        const next = root.k1.getElementByIndex!(0);
+        const item = root.k1.getElementByIndex!(2);
+        root.k1.moveBefore!(next.getID!(), item.getID!());
         assert.equal('{"k1":[2,0,1]}', root.toJSON!());
       });
 
       d1.update((root) => {
-        const next = root['k1'].getElementByIndex!(0);
-        const item = root['k1'].getElementByIndex!(2);
-        root['k1'].moveBefore!(next.getID(), item.getID());
+        const next = root.k1.getElementByIndex!(0);
+        const item = root.k1.getElementByIndex!(2);
+        root.k1.moveBefore!(next.getID!(), item.getID!());
         assert.equal('{"k1":[1,2,0]}', root.toJSON!());
       });
 
       d2.update((root) => {
-        const next = root['k1'].getElementByIndex!(0);
-        const item = root['k1'].getElementByIndex!(1);
-        root['k1'].moveBefore!(next.getID(), item.getID());
+        const next = root.k1.getElementByIndex!(0);
+        const item = root.k1.getElementByIndex!(1);
+        root.k1.moveBefore!(next.getID!(), item.getID!());
         assert.equal('{"k1":[1,0,2]}', root.toJSON!());
       });
 
       d2.update((root) => {
-        const next = root['k1'].getElementByIndex!(0);
-        const item = root['k1'].getElementByIndex!(1);
-        root['k1'].moveBefore!(next.getID(), item.getID());
+        const next = root.k1.getElementByIndex!(0);
+        const item = root.k1.getElementByIndex!(1);
+        root.k1.moveBefore!(next.getID!(), item.getID!());
         assert.equal('{"k1":[0,1,2]}', root.toJSON!());
       });
 
@@ -342,7 +347,7 @@ describe('Array', function () {
     type TestDoc = { k1: JSONArray<number> };
     await withTwoClientsAndDocuments<TestDoc>(async (c1, d1, c2, d2) => {
       d1.update((root) => {
-        root['k1'] = [0, 1, 2];
+        root.k1 = [0, 1, 2];
         assert.equal('{"k1":[0,1,2]}', root.toJSON!());
       });
       await c1.sync();
@@ -350,16 +355,16 @@ describe('Array', function () {
       assert.equal(d1.toJSON!(), d2.toJSON!());
 
       d1.update((root) => {
-        const next = root['k1'].getElementByIndex!(0);
-        const item = root['k1'].getElementByIndex!(1);
-        root['k1'].moveBefore!(next.getID(), item.getID());
+        const next = root.k1.getElementByIndex!(0);
+        const item = root.k1.getElementByIndex!(1);
+        root.k1.moveBefore!(next.getID!(), item.getID!());
         assert.equal('{"k1":[1,0,2]}', root.toJSON!());
       });
 
       d2.update((root) => {
-        const next = root['k1'].getElementByIndex!(1);
-        const item = root['k1'].getElementByIndex!(2);
-        root['k1'].moveBefore!(next.getID(), item.getID());
+        const next = root.k1.getElementByIndex!(1);
+        const item = root.k1.getElementByIndex!(2);
+        root.k1.moveBefore!(next.getID!(), item.getID!());
         assert.equal('{"k1":[0,2,1]}', root.toJSON!());
       });
 
@@ -374,7 +379,7 @@ describe('Array', function () {
     type TestDoc = { k1: JSONArray<number> };
     await withTwoClientsAndDocuments<TestDoc>(async (c1, d1, c2, d2) => {
       d1.update((root) => {
-        root['k1'] = [0, 1, 2];
+        root.k1 = [0, 1, 2];
         assert.equal('{"k1":[0,1,2]}', root.toJSON!());
       });
       await c1.sync();
@@ -382,26 +387,26 @@ describe('Array', function () {
       assert.equal(d1.toJSON!(), d2.toJSON!());
 
       d1.update((root) => {
-        const item = root['k1'].getElementByIndex!(2);
-        root['k1'].moveFront!(item.getID());
+        const item = root.k1.getElementByIndex!(2);
+        root.k1.moveFront!(item.getID!());
         assert.equal('{"k1":[2,0,1]}', root.toJSON!());
       });
 
       d1.update((root) => {
-        const item = root['k1'].getElementByIndex!(2);
-        root['k1'].moveFront!(item.getID());
+        const item = root.k1.getElementByIndex!(2);
+        root.k1.moveFront!(item.getID!());
         assert.equal('{"k1":[1,2,0]}', root.toJSON!());
       });
 
       d2.update((root) => {
-        const item = root['k1'].getElementByIndex!(1);
-        root['k1'].moveFront!(item.getID());
+        const item = root.k1.getElementByIndex!(1);
+        root.k1.moveFront!(item.getID!());
         assert.equal('{"k1":[1,0,2]}', root.toJSON!());
       });
 
       d2.update((root) => {
-        const item = root['k1'].getElementByIndex!(1);
-        root['k1'].moveFront!(item.getID());
+        const item = root.k1.getElementByIndex!(1);
+        root.k1.moveFront!(item.getID!());
         assert.equal('{"k1":[0,1,2]}', root.toJSON!());
       });
 
@@ -416,7 +421,7 @@ describe('Array', function () {
     type TestDoc = { k1: JSONArray<number> };
     await withTwoClientsAndDocuments<TestDoc>(async (c1, d1, c2, d2) => {
       d1.update((root) => {
-        root['k1'] = [0, 1, 2];
+        root.k1 = [0, 1, 2];
         assert.equal('{"k1":[0,1,2]}', root.toJSON!());
       });
       await c1.sync();
@@ -424,14 +429,14 @@ describe('Array', function () {
       assert.equal(d1.toJSON!(), d2.toJSON!());
 
       d1.update((root) => {
-        const item = root['k1'].getElementByIndex!(2);
-        root['k1'].moveFront!(item.getID());
+        const item = root.k1.getElementByIndex!(2);
+        root.k1.moveFront!(item.getID!());
         assert.equal('{"k1":[2,0,1]}', root.toJSON!());
       });
 
       d2.update((root) => {
-        const item = root['k1'].getElementByIndex!(2);
-        root['k1'].moveFront!(item.getID());
+        const item = root.k1.getElementByIndex!(2);
+        root.k1.moveFront!(item.getID!());
         assert.equal('{"k1":[2,0,1]}', root.toJSON!());
       });
 
@@ -446,7 +451,7 @@ describe('Array', function () {
     type TestDoc = { k1: JSONArray<number> };
     await withTwoClientsAndDocuments<TestDoc>(async (c1, d1, c2, d2) => {
       d1.update((root) => {
-        root['k1'] = [0, 1, 2];
+        root.k1 = [0, 1, 2];
         assert.equal('{"k1":[0,1,2]}', root.toJSON!());
       });
       await c1.sync();
@@ -454,14 +459,14 @@ describe('Array', function () {
       assert.equal(d1.toJSON!(), d2.toJSON!());
 
       d1.update((root) => {
-        const item = root['k1'].getElementByIndex!(1);
-        root['k1'].moveLast!(item.getID());
+        const item = root.k1.getElementByIndex!(1);
+        root.k1.moveLast!(item.getID!());
         assert.equal('{"k1":[0,2,1]}', root.toJSON!());
       });
 
       d2.update((root) => {
-        const item = root['k1'].getElementByIndex!(0);
-        root['k1'].moveLast!(item.getID());
+        const item = root.k1.getElementByIndex!(0);
+        root.k1.moveLast!(item.getID!());
         assert.equal('{"k1":[1,2,0]}', root.toJSON!());
       });
 
@@ -475,22 +480,22 @@ describe('Array', function () {
   it('Can handle concurrent insertAfter and moveBefore operations', async function () {
     type TestDoc = { k1: JSONArray<number> };
     await withTwoClientsAndDocuments<TestDoc>(async (c1, d1, c2, d2) => {
-      let prev: JSONElement;
+      let prev: WrappedElement;
       d1.update((root) => {
-        root['k1'] = [0];
-        prev = root['k1'].getElementByIndex!(0);
+        root.k1 = [0];
+        prev = root.k1.getElementByIndex!(0);
       });
       await c1.sync();
       await c2.sync();
       assert.equal(d1.toSortedJSON(), d2.toSortedJSON());
 
       d1.update((root) => {
-        root['k1'].insertAfter!(prev.getID(), 1);
+        root.k1.insertAfter!(prev.getID!(), 1);
         assert.equal('{"k1":[0,1]}', root.toJSON!());
       });
 
       d2.update((root) => {
-        root['k1'].insertAfter!(prev.getID(), 2);
+        root.k1.insertAfter!(prev.getID!(), 2);
         assert.equal('{"k1":[0,2]}', root.toJSON!());
       });
 
@@ -500,15 +505,15 @@ describe('Array', function () {
       assert.equal(d1.toSortedJSON(), d2.toSortedJSON());
 
       d1.update((root) => {
-        const next = root['k1'].getElementByIndex!(0);
-        const item = root['k1'].getElementByIndex!(1);
-        root['k1'].moveBefore!(next.getID(), item.getID());
+        const next = root.k1.getElementByIndex!(0);
+        const item = root.k1.getElementByIndex!(1);
+        root.k1.moveBefore!(next.getID!(), item.getID!());
       });
 
       d2.update((root) => {
-        const next = root['k1'].getElementByIndex!(0);
-        const item = root['k1'].getElementByIndex!(2);
-        root['k1'].moveBefore!(next.getID(), item.getID());
+        const next = root.k1.getElementByIndex!(0);
+        const item = root.k1.getElementByIndex!(2);
+        root.k1.moveBefore!(next.getID!(), item.getID!());
       });
 
       await c1.sync();
@@ -522,7 +527,7 @@ describe('Array', function () {
     type TestDoc = { k1: JSONArray<number> };
     await withTwoClientsAndDocuments<TestDoc>(async (c1, d1, c2, d2) => {
       d1.update((root) => {
-        root['k1'] = [0, 1, 2];
+        root.k1 = [0, 1, 2];
         assert.equal('{"k1":[0,1,2]}', root.toJSON!());
       });
       await c1.sync();
@@ -530,16 +535,16 @@ describe('Array', function () {
       assert.equal(d1.toJSON!(), d2.toJSON!());
 
       d1.update((root) => {
-        const prev = root['k1'].getElementByIndex!(0);
-        const item = root['k1'].getElementByIndex!(1);
-        root['k1'].moveAfter!(prev.getID(), item.getID());
+        const prev = root.k1.getElementByIndex!(0);
+        const item = root.k1.getElementByIndex!(1);
+        root.k1.moveAfter!(prev.getID!(), item.getID!());
         assert.equal('{"k1":[0,1,2]}', root.toJSON!());
       });
 
       d2.update((root) => {
-        const prev = root['k1'].getElementByIndex!(0);
-        const item = root['k1'].getElementByIndex!(2);
-        root['k1'].moveAfter!(prev.getID(), item.getID());
+        const prev = root.k1.getElementByIndex!(0);
+        const item = root.k1.getElementByIndex!(2);
+        root.k1.moveAfter!(prev.getID!(), item.getID!());
         assert.equal('{"k1":[0,2,1]}', root.toJSON!());
       });
 
@@ -554,17 +559,17 @@ describe('Array', function () {
     type TestDoc = { k1: Array<string> };
     await withTwoClientsAndDocuments<TestDoc>(async (c1, d1, c2, d2) => {
       d1.update((root) => {
-        root['k1'] = ['1'];
+        root.k1 = ['1'];
       });
       await c1.sync();
       await c2.sync();
       assert.equal(d1.toSortedJSON(), d2.toSortedJSON());
 
       d1.update((root) => {
-        root['k1'].push('2');
+        root.k1.push('2');
       });
       d2.update((root) => {
-        root['k1'].push('3');
+        root.k1.push('3');
       });
       await c1.sync();
       await c2.sync();
@@ -576,29 +581,29 @@ describe('Array', function () {
   it('Can handle concurrent delete operations', async function () {
     type TestDoc = { k1: JSONArray<number> };
     await withTwoClientsAndDocuments<TestDoc>(async (c1, d1, c2, d2) => {
-      let prev: JSONElement;
+      let prev: WrappedElement;
       d1.update((root) => {
-        root['k1'] = [1, 2, 3, 4];
+        root.k1 = [1, 2, 3, 4];
       });
       await c1.sync();
       await c2.sync();
       assert.equal(d1.toSortedJSON(), d2.toSortedJSON());
 
       d1.update((root) => {
-        prev = root['k1'].getElementByIndex!(2);
-        root['k1'].deleteByID!(prev.getCreatedAt());
+        prev = root.k1.getElementByIndex!(2);
+        root.k1.deleteByID!(prev.getID!());
       });
 
       d2.update((root) => {
-        prev = root['k1'].getElementByIndex!(2);
-        root['k1'].deleteByID!(prev.getCreatedAt());
+        prev = root.k1.getElementByIndex!(2);
+        root.k1.deleteByID!(prev.getID!());
       });
 
       await c1.sync();
       await c2.sync();
       await c1.sync();
       d1.update((root) => {
-        assert.equal(3, root['k1'].length);
+        assert.equal(3, root.k1.length);
       });
     }, this.test!.title);
   });
@@ -606,25 +611,25 @@ describe('Array', function () {
   it('Can handle concurrent insertBefore and delete operations', async function () {
     type TestDoc = { k1: JSONArray<number> };
     await withTwoClientsAndDocuments<TestDoc>(async (c1, d1, c2, d2) => {
-      let prev: JSONElement;
+      let prev: WrappedElement;
 
       d1.update((root) => {
-        root['k1'] = [1];
-        prev = root['k1'].getElementByIndex!(0);
+        root.k1 = [1];
+        prev = root.k1.getElementByIndex!(0);
       });
       await c1.sync();
       await c2.sync();
       assert.equal(d1.toSortedJSON(), d2.toSortedJSON());
 
       d1.update((root) => {
-        root['k1'].deleteByID!(prev.getID());
+        root.k1.deleteByID!(prev.getID!());
         assert.equal('{"k1":[]}', root.toJSON!());
-        assert.equal(0, root['k1'].length);
+        assert.equal(0, root.k1.length);
       });
       d2.update((root) => {
-        root['k1'].insertBefore!(prev.getID(), 2);
+        root.k1.insertBefore!(prev.getID!(), 2);
         assert.equal('{"k1":[2,1]}', root.toJSON!());
-        assert.equal(2, root['k1'].length);
+        assert.equal(2, root.k1.length);
       });
       await c1.sync();
       await c2.sync();
@@ -632,7 +637,7 @@ describe('Array', function () {
       assert.equal(d1.toSortedJSON(), d2.toSortedJSON());
       d1.update((root) => {
         assert.equal('{"k1":[2]}', root.toJSON!());
-        assert.equal(1, root['k1'].length);
+        assert.equal(1, root.k1.length);
       });
     }, this.test!.title);
   });
@@ -640,25 +645,25 @@ describe('Array', function () {
   it('Can handle complex concurrent insertBefore and delete operations', async function () {
     type TestDoc = { k1: JSONArray<number> };
     await withTwoClientsAndDocuments<TestDoc>(async (c1, d1, c2, d2) => {
-      let prev: JSONElement;
+      let prev: WrappedElement;
 
       d1.update((root) => {
-        root['k1'] = [1, 2, 3, 4];
-        prev = root['k1'].getElementByIndex!(1);
+        root.k1 = [1, 2, 3, 4];
+        prev = root.k1.getElementByIndex!(1);
       });
       await c1.sync();
       await c2.sync();
       assert.equal(d1.toSortedJSON(), d2.toSortedJSON());
 
       d1.update((root) => {
-        root['k1'].deleteByID!(prev.getID());
+        root.k1.deleteByID!(prev.getID!());
         assert.equal('{"k1":[1,3,4]}', root.toJSON!());
-        assert.equal(3, root['k1'].length);
+        assert.equal(3, root.k1.length);
       });
       d2.update((root) => {
-        root['k1'].insertBefore!(prev.getID(), 5);
+        root.k1.insertBefore!(prev.getID!(), 5);
         assert.equal('{"k1":[1,5,2,3,4]}', root.toJSON!());
-        assert.equal(5, root['k1'].length);
+        assert.equal(5, root.k1.length);
       });
       await c1.sync();
       await c2.sync();
@@ -667,18 +672,18 @@ describe('Array', function () {
       assert.equal('{"k1":[1,5,3,4]}', d1.toJSON!());
 
       d1.update((root) => {
-        const prev = root['k1'].getElementByIndex!(3);
-        assert.equal(4, root['k1'].length);
-        root['k1'].insertBefore!(prev.getID(), 6);
+        const prev = root.k1.getElementByIndex!(3);
+        assert.equal(4, root.k1.length);
+        root.k1.insertBefore!(prev.getID!(), 6);
         assert.equal('{"k1":[1,5,3,6,4]}', root.toJSON!());
-        assert.equal(5, root['k1'].length);
+        assert.equal(5, root.k1.length);
       });
       d2.update((root) => {
-        const prev = root['k1'].getElementByIndex!(0);
-        assert.equal(4, root['k1'].length);
-        root['k1'].insertBefore!(prev.getID(), 7);
+        const prev = root.k1.getElementByIndex!(0);
+        assert.equal(4, root.k1.length);
+        root.k1.insertBefore!(prev.getID!(), 7);
         assert.equal('{"k1":[7,1,5,3,4]}', root.toJSON!());
-        assert.equal(5, root['k1'].length);
+        assert.equal(5, root.k1.length);
       });
       await c1.sync();
       await c2.sync();
@@ -686,7 +691,7 @@ describe('Array', function () {
       assert.equal(d1.toSortedJSON(), d2.toSortedJSON());
       d1.update((root) => {
         assert.equal('{"k1":[7,1,5,3,6,4]}', root.toJSON!());
-        assert.equal(6, root['k1'].length);
+        assert.equal(6, root.k1.length);
       });
     }, this.test!.title);
   });
@@ -696,19 +701,19 @@ describe('Array', function () {
     let targetID: TimeTicket;
 
     doc.update((root) => {
-      root['list'] = [0, 1, 2];
-      targetID = root['list'].getElementByIndex!(2).getID();
+      root.list = [0, 1, 2];
+      targetID = root.list.getElementByIndex!(2).getID!();
     });
 
     doc.update((root) => {
-      root['list'].deleteByID!(targetID);
+      root.list.deleteByID!(targetID);
     });
 
     assert.equal('{"list":[0,1]}', doc.toSortedJSON());
 
     doc.garbageCollect(MaxTimeTicket);
     doc.update((root) => {
-      const elem = root['list'].getElementByID!(targetID);
+      const elem = root.list.getElementByID!(targetID);
       assert.isUndefined(elem);
     });
   });
@@ -718,17 +723,18 @@ describe('Array', function () {
     let targetID: TimeTicket;
 
     doc.update((root) => {
-      root['list'] = [0, 1, 2];
-      targetID = root['list'].getElementByIndex!(2).getID();
+      root.list = [0, 1, 2];
+      targetID = root.list.getElementByIndex!(2).getID!();
     });
 
     doc.update((root) => {
-      assert.equal(2, root['list'].getElementByID!(targetID));
+      const elem = root.list.getElementByID!(targetID) as Primitive;
+      assert.equal(2, elem.getValue());
     });
 
     doc.update((root) => {
-      root['list'].deleteByID!(targetID);
-      assert.isUndefined(root['list'].getElementByID!(targetID));
+      root.list.deleteByID!(targetID);
+      assert.isUndefined(root.list.getElementByID!(targetID));
     });
 
     assert.equal('{"list":[0,1]}', doc.toSortedJSON());
@@ -738,7 +744,7 @@ describe('Array', function () {
     type TestDoc = {
       empty: [];
       list: JSONArray<number>;
-      listObjects: JSONArray<{ id: string }>;
+      objects: JSONArray<{ id: string }>;
     };
 
     it('concat()', () => {
@@ -847,7 +853,7 @@ describe('Array', function () {
       doc.update((root) => {
         root.empty = [];
         root.list = [1, 2, 3];
-        root.listObjects = [{ id: 'first' }, { id: 'second' }];
+        root.objects = [{ id: 'first' }, { id: 'second' }];
       });
 
       assert.strictEqual(
@@ -866,7 +872,7 @@ describe('Array', function () {
       );
 
       assert.deepEqual(
-        doc.getRoot().listObjects.find((obj) => obj.id === 'first'),
+        doc.getRoot().objects.find((obj) => obj.id === 'first'),
         { id: 'first' },
       );
 
@@ -963,12 +969,12 @@ describe('Array', function () {
     it('indexOf() with objects', () => {
       const doc = DocumentReplica.create<TestDoc>('test-doc');
       doc.update((root) => {
-        root.listObjects = [{ id: 'first' }, { id: 'second' }];
+        root.objects = [{ id: 'first' }, { id: 'second' }];
       });
 
       // TODO: test always fails because doc.getRoot() returns a new proxy of cloned root.
       // assert.strictEqual(
-      //   doc.getRoot().listObjects.indexOf(doc.getRoot().listObjects[0]),
+      //   doc.getRoot().objects.indexOf(doc.getRoot().objects[0]),
       //   0,
       // );
     });
@@ -1171,7 +1177,7 @@ describe('Array', function () {
       doc.update((root) => {
         root.empty = [];
         root.list = [1, 2, 3];
-        root.listObjects = [{ id: '1' }, { id: '2' }, { id: '3' }];
+        root.objects = [{ id: '1' }, { id: '2' }, { id: '3' }];
       });
 
       assert.strictEqual(doc.getRoot().empty.toString(), '');
@@ -1180,7 +1186,7 @@ describe('Array', function () {
       // NOTE: This is not the same as the listObjects.toString()
       //       "[object object],[object object],[object object]"
       assert.strictEqual(
-        doc.getRoot().listObjects.toString(),
+        doc.getRoot().objects.toString(),
         '{"id":"1"},{"id":"2"},{"id":"3"}',
       );
     });
@@ -1202,11 +1208,11 @@ describe('Array', function () {
     it('should allow mutation of objects returned from built in list iteration', () => {
       const doc = DocumentReplica.create<TestDoc>('test-doc');
       doc.update((root) => {
-        root.listObjects = [{ id: 'first' }, { id: 'second' }];
+        root.objects = [{ id: 'first' }, { id: 'second' }];
       });
 
       doc.update((root) => {
-        for (const obj of root.listObjects) {
+        for (const obj of root.objects) {
           if (obj.id === 'first') {
             obj.id = 'FIRST';
           }
@@ -1215,23 +1221,23 @@ describe('Array', function () {
 
       assert.equal(
         doc.toSortedJSON(),
-        '{"listObjects":[{"id":"FIRST"},{"id":"second"}]}',
+        '{"objects":[{"id":"FIRST"},{"id":"second"}]}',
       );
     });
 
     it('should allow mutation of objects returned from readonly list methods', () => {
       const doc = DocumentReplica.create<TestDoc>('test-doc');
       doc.update((root) => {
-        root.listObjects = [{ id: 'first' }, { id: 'second' }];
+        root.objects = [{ id: 'first' }, { id: 'second' }];
       });
 
       doc.update((root) => {
-        root.listObjects.find((obj) => obj.id === 'first')!.id = 'FIRST';
+        root.objects.find((obj) => obj.id === 'first')!.id = 'FIRST';
       });
 
       assert.equal(
         doc.toSortedJSON(),
-        '{"listObjects":[{"id":"FIRST"},{"id":"second"}]}',
+        '{"objects":[{"id":"FIRST"},{"id":"second"}]}',
       );
     });
   });
