+++ test/unit/document/json/root_test.ts
@@ -1,28 +1,28 @@
 import { assert } from 'chai';
 import { InitialChangeID } from '@yorkie-js-sdk/src/document/change/change_id';
-import { JSONRoot } from '@yorkie-js-sdk/src/document/json/root';
-import { ObjectInternal } from '@yorkie-js-sdk/src/document/json/object';
-import { RHTPQMap } from '@yorkie-js-sdk/src/document/json/rht_pq_map';
+import { CRDTRoot } from '@yorkie-js-sdk/src/document/crdt/root';
+import { CRDTObject } from '@yorkie-js-sdk/src/document/crdt/object';
+import { RHTPQMap } from '@yorkie-js-sdk/src/document/crdt/rht_pq_map';
 import { ChangeContext } from '@yorkie-js-sdk/src/document/change/context';
-import { ObjectProxy } from '@yorkie-js-sdk/src/document/proxy/object_proxy';
-import { ArrayProxy } from '@yorkie-js-sdk/src/document/proxy/array_proxy';
+import { ObjectProxy } from '@yorkie-js-sdk/src/document/json/object';
+import { ArrayProxy } from '@yorkie-js-sdk/src/document/json/array';
 import { InitialTimeTicket } from '@yorkie-js-sdk/src/document/time/ticket';
 import { MaxTimeTicket } from '@yorkie-js-sdk/src/document/time/ticket';
-import { RGATreeList } from '@yorkie-js-sdk/src/document/json/rga_tree_list';
-import { JSONPrimitive } from '@yorkie-js-sdk/src/document/json/primitive';
-import { ArrayInternal } from '@yorkie-js-sdk/src/document/json/array';
+import { RGATreeList } from '@yorkie-js-sdk/src/document/crdt/rga_tree_list';
+import { Primitive } from '@yorkie-js-sdk/src/document/crdt/primitive';
+import { CRDTArray } from '@yorkie-js-sdk/src/document/crdt/array';
 
 describe('ROOT', function () {
   it('basic test', function () {
-    const root = new JSONRoot(
-      new ObjectInternal(InitialTimeTicket, RHTPQMap.create()),
+    const root = new CRDTRoot(
+      new CRDTObject(InitialTimeTicket, RHTPQMap.create()),
     );
     const cc = ChangeContext.create(InitialChangeID, root);
     assert.isUndefined(root.findByCreatedAt(MaxTimeTicket));
     assert.isUndefined(root.createPath(MaxTimeTicket));
 
     // set '$.k1'
-    const k1 = JSONPrimitive.of('k1', cc.issueTimeTicket());
+    const k1 = Primitive.of('k1', cc.issueTimeTicket());
     root.getObject().set('k1', k1);
     root.registerElement(k1, root.getObject());
     assert.equal(root.getElementMapSize(), 2);
@@ -37,7 +37,7 @@ describe('ROOT', function () {
     assert.isUndefined(root.findByCreatedAt(k1.getCreatedAt()));
 
     // set '$.k2'
-    const k2 = ObjectInternal.create(cc.issueTimeTicket());
+    const k2 = CRDTObject.create(cc.issueTimeTicket());
     root.getObject().set('k2', k2);
     root.registerElement(k2, root.getObject());
     assert.equal(root.getElementMapSize(), 2);
@@ -47,7 +47,7 @@ describe('ROOT', function () {
     assert.equal(Object.keys(k2.toJS()).length, 0);
 
     // set '$.k2.1'
-    const k2_1 = ArrayInternal.create(cc.issueTimeTicket());
+    const k2_1 = CRDTArray.create(cc.issueTimeTicket());
     k2.set('1', k2_1);
     root.registerElement(k2_1, k2);
     assert.equal(root.getElementMapSize(), 3);
@@ -55,7 +55,7 @@ describe('ROOT', function () {
     assert.equal(root.createPath(k2_1.getCreatedAt()), '$.k2.1');
 
     // set '$.k2.1.0'
-    const k2_1_0 = JSONPrimitive.of('0', cc.issueTimeTicket());
+    const k2_1_0 = Primitive.of('0', cc.issueTimeTicket());
     k2_1.insertAfter(k2_1.getLastCreatedAt(), k2_1_0);
     root.registerElement(k2_1_0, k2_1);
     assert.equal(root.getElementMapSize(), 4);
@@ -63,7 +63,7 @@ describe('ROOT', function () {
     assert.equal(root.createPath(k2_1_0.getCreatedAt()), '$.k2.1.0');
 
     // set '$.k2.1.1'
-    const k2_1_1 = JSONPrimitive.of('1', cc.issueTimeTicket());
+    const k2_1_1 = Primitive.of('1', cc.issueTimeTicket());
     k2_1.insertAfter(k2_1_0.getCreatedAt(), k2_1_1);
     root.registerElement(k2_1_1, k2_1);
     assert.equal(root.getElementMapSize(), 5);
@@ -72,10 +72,10 @@ describe('ROOT', function () {
   });
 
   it('garbage collection test for array', function () {
-    const root = new JSONRoot(
-      new ObjectInternal(InitialTimeTicket, RHTPQMap.create()),
+    const root = new CRDTRoot(
+      new CRDTObject(InitialTimeTicket, RHTPQMap.create()),
     );
-    const arr = new ArrayInternal(InitialTimeTicket, RGATreeList.create());
+    const arr = new CRDTArray(InitialTimeTicket, RGATreeList.create());
     const change = ChangeContext.create(InitialChangeID, root);
 
     ArrayProxy.pushInternal(change, arr, 0);
@@ -103,10 +103,10 @@ describe('ROOT', function () {
   });
 
   it('garbage collection test for text', function () {
-    const root = new JSONRoot(
-      new ObjectInternal(InitialTimeTicket, RHTPQMap.create()),
+    const root = new CRDTRoot(
+      new CRDTObject(InitialTimeTicket, RHTPQMap.create()),
     );
-    const obj = new ObjectInternal(InitialTimeTicket, RHTPQMap.create());
+    const obj = new CRDTObject(InitialTimeTicket, RHTPQMap.create());
     const change = ChangeContext.create(InitialChangeID, root);
     const text = ObjectProxy.createText(change, obj, 'k1');
 
@@ -125,10 +125,10 @@ describe('ROOT', function () {
   });
 
   it('garbage collection test for rich text', function () {
-    const root = new JSONRoot(
-      new ObjectInternal(InitialTimeTicket, RHTPQMap.create()),
+    const root = new CRDTRoot(
+      new CRDTObject(InitialTimeTicket, RHTPQMap.create()),
     );
-    const obj = new ObjectInternal(InitialTimeTicket, RHTPQMap.create());
+    const obj = new CRDTObject(InitialTimeTicket, RHTPQMap.create());
     const change = ChangeContext.create(InitialChangeID, root);
     const text = ObjectProxy.createRichText(change, obj, 'k1');
